+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Development]
 
+## [dev0.74] - 2019-04-19
+
 ### Changed
 - `DataStructure.Generic.Recording.InputPath` becomes `GetInputPath` for clarity and consistency
 
@@ -12,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - library for BIDS fields used in .tsv files. New class `tools.json.fieldEntry` is defined with a list of field definitions. New fields can be added via AddField. `DumpDescription` creates a descriptive json file, while GetHeader and GetLine generates the header and individual lines for .tsv files
 - fieldEntry is defined for participants.tsv and events.tsv files
 - scan id functions and fieldLibrary moved to BIDS/BIDS.py
+- `-p --patient` and `[GENERAL]PatientId` options to manually set the id of subject. It follows usual priority rule: file < cfi < cli
 
 ### Fixed
 - Sidecar Json created without run label